Move read-secret command to a different package per review feedback
