+++ cmd/secrets/testdata/read-secrets/secret3
